requirements.txt
aioredis==1.3.1
apache-libcloud==3.5.1
argcomplete==2.0.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
boto3==1.23.10
botocore==1.26.10
certifi==2021.10.8
cffi==1.15.0
channels==2.4.0
channels-redis==2.4.2
charset-normalizer==2.0.9
click==8.1.3
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==2.5.0
defusedxml==0.7.1
distlib==0.3.4
Django==3.2.9
django-cors-headers==3.10.1
django-extensions==3.1.5
django-filter==21.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Faker==13.3.4
filelock==3.4.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.0
MarkupSafe==2.0.1
msgpack==0.6.2
mysqlclient==2.1.0
oauthlib==3.1.1
packaging==23.0
Pillow==8.4.0
pipenv==2022.12.19
pipx==1.1.0
platformdirs==2.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
userpath==1.8.0
virtualenv==20.13.0
virtualenv-clone==0.5.7
zope.interface==5.4.0
